+++ src/git/abbreviation.rs
@@ -11,3 +11,25 @@ pub fn abbreviated_git_url_to_full_remote(git: impl AsRef<str>) -> String {
         git.to_owned()
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn should_support_bb_gl_gh_abbreviations() {
+        assert_eq!(
+            &abbreviated_git_url_to_full_remote("gh:foo/bar"),
+            "https://github.com/foo/bar.git"
+        );
+        assert_eq!(
+            &abbreviated_git_url_to_full_remote("bb:foo/bar"),
+            "https://bitbucket.org/foo/bar.git"
+        );
+        assert_eq!(
+            &abbreviated_git_url_to_full_remote("gl:foo/bar"),
+            "https://gitlab.com/foo/bar.git"
+        );
+        assert_eq!(&abbreviated_git_url_to_full_remote("foo/bar"), "foo/bar");
+    }
+}
